refactor: move Offers database to its own project
