+++ Chapter-2-modules-separation/Src/Fitnet.Offers.DataAccess/DataAccessModule.cs
@@ -1,0 +1,23 @@
+namespace EvolutionaryArchitecture.Fitnet.Offers.DataAccess;
+
+using Database;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+public static class DataAccessModule
+{
+    public static IServiceCollection AddDataAccess(this IServiceCollection services, IConfiguration configuration)
+    {
+        services.AddDatabase(configuration);
+
+        return services;
+    }
+    
+    public static IApplicationBuilder UseDataAccess(this IApplicationBuilder applicationBuilder)
+    {
+        applicationBuilder.UseDatabase();
+
+        return applicationBuilder;
+    }
+}
