+++ Chapter-2-modules-separation/Src/Fitnet.Offers.DataAccess/Database/AutomaticMigrationsExtensions.cs
@@ -1,6 +1,8 @@
-namespace EvolutionaryArchitecture.Fitnet.Offers.Data.Database;
+namespace EvolutionaryArchitecture.Fitnet.Offers.DataAccess.Database;
 
+using Microsoft.AspNetCore.Builder;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.DependencyInjection;
 
 internal static class AutomaticMigrationsExtensions
 {
