+++ Chapter-2-modules-separation/Src/Fitnet.Offers.DataAccess/Database/DatabaseModule.cs
@@ -1,0 +1,26 @@
+namespace EvolutionaryArchitecture.Fitnet.Offers.DataAccess.Database;
+
+using Microsoft.AspNetCore.Builder;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+public static class DatabaseModule
+{
+    private const string ConnectionStringName = "Offers";
+
+    public static IServiceCollection AddDatabase(this IServiceCollection services, IConfiguration configuration)
+    {
+        var connectionString = configuration.GetConnectionString(ConnectionStringName);
+        services.AddDbContext<OffersPersistence>(options => options.UseNpgsql(connectionString));
+
+        return services;
+    }
+
+    public static IApplicationBuilder UseDatabase(this IApplicationBuilder applicationBuilder)
+    {
+        applicationBuilder.UseAutomaticMigrations();
+
+        return applicationBuilder;
+    }
+}
