+++ Chapter-2-modules-separation/Src/Fitnet.Offers.DataAccess/GlobalUsings.cs
@@ -1,0 +1,2 @@
+global using System;
+global using System.Diagnostics.CodeAnalysis;
