--- Chapter-2-modules-separation/Src/Fitnet/Offers/Data/Database/DatabaseModule.cs
@@ -1,23 +1,0 @@
-namespace EvolutionaryArchitecture.Fitnet.Offers.Data.Database;
-
-using Microsoft.EntityFrameworkCore;
-
-internal static class DatabaseModule
-{
-    private const string ConnectionStringName = "Offers";
-
-    internal static IServiceCollection AddDatabase(this IServiceCollection services, IConfiguration configuration)
-    {
-        var connectionString = configuration.GetConnectionString(ConnectionStringName);
-        services.AddDbContext<OffersPersistence>(options => options.UseNpgsql(connectionString));
-
-        return services;
-    }
-
-    internal static IApplicationBuilder UseDatabase(this IApplicationBuilder applicationBuilder)
-    {
-        applicationBuilder.UseAutomaticMigrations();
-
-        return applicationBuilder;
-    }
-}
